TD/js-ast-toolkit/src/backend/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.30
Flask-SocketIO==5.3.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0